Species has primary key 'id' instead of using sPrimeID
Not everything has a prime ID, and we may end up with one species
having two prime IDs, or crazy things. Eventually both databases
should be one-species-per-ID, but for now we probably disagree
what comprises a "Species".

This migration introduces a new autoincrement field, so will not
work on a database already containing data.