Listed every xml component from PrIMe warehouse
The description now contains hierarchically organized lists of all fields that could be necessary in the new django database when importing PrIMe data into it. Need to establish which parameters are worth preserving from PrIMe and how many django classes should be created and overlapped to accommodate the vast amounts of categories.