Made an import_prime.py script to import prime to django
You run this, give it the path to your local prime warehouse 
mirror, and it imports everything.
So far only bibliography is implemented, and not all fields there,
but it gives you a good idea how to go about doing this.
Should be able to flesh it out and add more fields and objects.