Assign species names to variables
The species names are assigned to changing variables (an incremented integer for each loop) via a dictionary. Hopefully this method of assignment will function well enough to print out the correct xml script at the end of the entire species class (it must all be printed at once and so could not simply be printed for each loop).