Added authors and moved files around
Attempting to get each type of prime ID in a separate folder so that each might be easily streamlined into PrIMe, but am currently getting errors. Four commented lines at the bottom of writertest.py are attempting this. However, succeeded in accessing authors via ForeignKey, and they are now printed in the correct order in the xml file.