Copied over import scripts and made some changes to rmg import script
Moved those files over and did some py2to3 changes for the rmg import script. Also added "errors.txt" to the ignore file. There is an issue where running once adds thermo but not the species info. Then running it again adds the species information but causes a duplicate of thermo to be added... working on just thermo at the moment.