Refactoring (#40)
* add black project configuration

black will now format with a line length of 100 and will exclude common
python project files along with django migration folders.

* run black

* remove pipenv config

* add dev dependencies

* add contributing section to README

* remove unused and old site code

Most of these templates and code are from the MIT site and are not
currently used in this stage of the project. They can be added back as
needed later on.

* organize apps and template settings

* remove unneeded notes

* remove unused views and properly pass context data

* consolidate, consistently inherit from base, and clean up templates

* fix incorrect detail lookup

* remove unused static files

* clean up .gitignore

* fix minor errors and typos

* simplify __init__

* convert fields from camelCase to snake_case

* remove pipenv instructions from README

* add version control tips to README

* remove unused app (?)

This folder is unused and it is not clear what it actually was at some
point given that is has a migration and a script in it.