Updating the Thermo class methods for enthlapy, entropy, etc.
Now you specify the temperature, and it picks the right polynomial
and evaluates the thing you ask for.
Also made some properties for the values at 298K, for rendering
in templates.

I removed the things that give a range of values, because I couldn't
see where they were being used.