search for reaction before creating and add species simultaneously
The import flow is now to collect the species-stoichiometry pairs and
attempt to find the reaction. If that fails, a new reaction is created
with the collected stoichiometries. Additionally, the stoichiometries
are all added at the same time before saving so that the set uniqueness validation
doesn't fail if a subset of stoichiometries is not unique.