fail if reaction has no product or no reactants
If something goes wrong and a reaction ends up with no reactants or no
products, the reaction import should fail. The Reaction.reactants() and
Reaction.products() methods are not used since they do not exist for
some reason in this data migration. The filters include stoichiometries
with a coefficient of 0 for species that are on both sides of the
reaction.