format stoich coefficient in hash fingerprint
This hashing function now provides the same hash whether the
stoichiometric coefficients are floats or integers, which caused
confusion and issues during testing since the coefficients are integers
in the importer script, but are stored as floats in the database. This
change hopefully also fixes precision issues with fractional
coefficients.