Create stoichiometry objects explicitly
It was previously thought that an object with a ManyToManyField that has
a 'through' object specified could have its members extended by using
the add(through_defaults=...) method. This did not work when adding the
same Species to a Reaction multiple times for cases where the Species is
on both sides of the Reaction. Instead, Stoichiometry objects are
explicity created for each Species when a Reaction is created.