add db signals to update hashes when m2m changes
This change attempts to keep Species and Reaction hashes synced
with their respective ManyToManyFields by using database signals. It is
unclear whether this is absolutely necessary, as exposure of only
creation and deletion forms for these objects and judicious updates of
Species and Reactions via deletion and recreation may suffice, but it
could save some headache.