add scripts for debugging in VSCode
Add two scripts for running the server and running migrations in the
debug docker container with the VSCode debugger. There is currently a
bug when using `docker-compose run` where either script runs and waits
for a client to attach, but when launching the debug configuration in
VSCode, the connection is refused. This does not happen when either
script is in the "command" definition in the docker-compose.debug file.