only run unit tests by default
The migration tests take a very long time to run, and should only be run
manually. Running the migration tests is not only time-consuming to run
in CI, but require dependencies such as the RMG-Models repo, making it
infeasible to run. All migration test files should be suffixed with
"mig" by convention.